requirements.txt
asgiref==3.2.7
certifi==2020.12.5
chardet==4.0.0
Django==3.0.7
django-crispy-forms==1.9.2
django-filter==2.3.0
idna==2.10
Pillow==7.1.2
psycopg2==2.8.6
pytz==2019.3
requests==2.25.1
sqlparse==0.3.0
urllib3==1.26.4